+++ Sources/Connect/GenericConnect.swift
@@ -63,7 +63,7 @@ fileprivate struct Default {
 /// GenericConnect is the concrete implementation of the Connect protocol.  Use this class to instantiate 
 /// a specifically configured instance for you needs.
 ///
-public class GenericConnect<Strategy: ContextStrategyType>: Connect {
+public class GenericConnect<Strategy: ContextStrategyType>: NSObject, Connect {
 
     ///
     /// Internal types defining the MetaCache and DataCache CoreDataStack types
@@ -72,9 +72,23 @@ public class GenericConnect<Strategy: ContextStrategyType>: Connect {
     fileprivate typealias MetaCacheType = GenericPersistentContainer<ContextStrategy.DirectIndependent>
 
     ///
+    /// Creates and returns a URL to the default directory for the persistent stores.
+    ///
+    public static func defaultStoreLocation() -> URL {
+        return Coherence.defaultStoreLocation()
+    }
+
+    ///
     /// The name of this instance of `Connect`
     ///
     public let name: String
+
+    ///
+    /// The model this instance was constructed with.
+    ///
+    public var managedObjectModel: NSManagedObjectModel {
+        return persistentStoreCoordinator.managedObjectModel
+    }
 
     ///
     /// Returns the `NSPersistentStoreCoordinate` instance that
@@ -200,6 +214,8 @@ public class GenericConnect<Strategy: ContextStrategyType>: Connect {
 
         self.started = false
 
+        super.init()
+
         ///
         /// Note: due to Swift requirement for not passing self until all instance
         ///       variables have a value, this has to be set here and not in the 
@@ -244,12 +260,8 @@ public class GenericConnect<Strategy: ContextStrategyType>: Connect {
             self._suspended = true
         }
     }
-}
 
-///
-/// Connect state management (public synchronized)
-///
-public extension GenericConnect {
+    // MARK: - Connect state management (public synchronized)
 
     ///
     /// Synchronously start the instance of `Connect`, starting all
@@ -333,6 +345,7 @@ public extension GenericConnect {
     /// - Note: Calling this method will attempt to create the directory specified in the `at` parameter specified.  If a url is not specified, the `defaultStoreLocation` will be used and an attempt will be made to create that directory.
     ///
     @discardableResult
+    @objc(attachPersistentStoreAt:for:error:)
     public func attachPersistentStore(at url: URL, for configuration: StoreConfiguration) throws -> NSPersistentStore {
         return try self.synchronizationQueue.sync {
             return try self._attachPersistentStore(at: url, for: configuration)
@@ -342,6 +355,7 @@ public extension GenericConnect {
     ///
     /// Detach a persistent store from the Coordinator.
     ///
+    @objc(detachWithPersistentStore:error:)
     public func detach(persistentStore store: NSPersistentStore) throws {
         try self.synchronizationQueue.sync {
             try self._detach(persistentStore: store)
@@ -359,6 +373,7 @@ public extension GenericConnect {
     /// - Note: Calling this method will attempt to create the directory specified in the `at` parameter specified.  If a url is not specified, the `defaultStoreLocation` will be used and an attempt will be made to create that directory.
     ///
     @discardableResult
+    @objc(attachPersistentStoresAt:for:error:)
     public func attachPersistentStores(at url: URL, for configurations: [StoreConfiguration]) throws -> [NSPersistentStore] {
 
         return try self.synchronizationQueue.sync {
@@ -374,6 +389,7 @@ public extension GenericConnect {
     ///
     /// Detach an array of persistent stores from the Coordinator.
     ///
+    @objc(detachWithPersistentStores:error:)
     public func detach(persistentStores stores: [NSPersistentStore]) throws {
         try self.synchronizationQueue.sync {
             for store in stores {
@@ -400,12 +416,8 @@ public extension GenericConnect {
             }
         }
     }
-}
 
-///
-/// Context access methods
-///
-extension GenericConnect {
+    // MARK: - Context access methods
 
     ///
     /// The main context.
@@ -414,6 +426,7 @@ extension GenericConnect {
     ///
     /// It will be maintained automatically and be kept consistent.
     ///
+    @objc
     public var viewContext: NSManagedObjectContext {
         return self.dataCache.viewContext
     }
@@ -423,6 +436,7 @@ extension GenericConnect {
     ///
     /// At save time, Connect will merge those changes back to the ViewContextType.
     ///
+    @objc
     public func newBackgroundContext() -> BackgroundContext {
         return self.newBackgroundContext(logged: true)
     }
@@ -436,7 +450,7 @@ extension GenericConnect {
     ///
     public func newBackgroundContext(logged: Bool) -> BackgroundContext {
 
-        let context: LoggingContext = self.dataCache.newBackgroundContext()
+        let context: LoggingContext = self.dataCache.newGenericBackgroundContext()
 
         if logged {
             ///
@@ -450,7 +464,7 @@ extension GenericConnect {
 
     internal func newActionContext() -> ActionContext {
 
-        let context: ActionContext = self.dataCache.newBackgroundContext()
+        let context: ActionContext = self.dataCache.newGenericBackgroundContext()
         
         context.mergePolicy = NSMergeByPropertyObjectTrumpMergePolicy
         context.logger = self.writeAheadLog
@@ -462,7 +476,7 @@ extension GenericConnect {
 ///
 /// Action Execution methods
 ///
-public extension GenericConnect {
+extension GenericConnect: ActionManager {
 
     ///
     /// Execute a generic action in a concurrent queue.
@@ -579,7 +593,7 @@ fileprivate extension GenericConnect {
 
             try self.metaCache.attachPersistentStore(at: GenericConnect.defaultStoreLocation(), for: metaStoreConfiguration)
 
-            self.writeAheadLog = try WriteAheadLog(persistentStack: self.metaCache)
+            self.writeAheadLog = try WriteAheadLog(persistentContainer: self.metaCache)
 
             ///
             /// If there are no stores attached at the time of starting, we assume that